+++ include/AudioOss.h
@@ -30,6 +30,7 @@
 #ifdef LMMS_HAVE_OSS
 
 #include "AudioDevice.h"
+#include "AudioDeviceSetupWidget.h"
 
 
 class LcdSpinBox;
@@ -50,7 +51,7 @@ public:
 	static QString probeDevice();
 
 
-	class setupWidget : public AudioDevice::setupWidget
+	class setupWidget : public AudioDeviceSetupWidget
 	{
 	public:
 		setupWidget( QWidget * _parent );
