+++ src/core/audio/AudioOss.cpp
@@ -329,7 +329,7 @@ void AudioOss::run()
 
 
 AudioOss::setupWidget::setupWidget( QWidget * _parent ) :
-	AudioDevice::setupWidget( AudioOss::name(), _parent )
+	AudioDeviceSetupWidget( AudioOss::name(), _parent )
 {
 	m_device = new QLineEdit( probeDevice(), this );
 	m_device->setGeometry( 10, 20, 160, 20 );
