+++ src/gui/CMakeLists.txt
@@ -2,6 +2,7 @@ SET(LMMS_SRCS
 	${LMMS_SRCS}
 	gui/AboutDialog.cpp
 	gui/ActionGroup.cpp
+	gui/AudioAlsaSetupWidget.cpp
 	gui/AutomatableModelView.cpp
 	gui/AutomationPatternView.cpp
 	gui/ControllerConnectionDialog.cpp
